// open-tactics-core/src/domain/ethnicity.enum.ts
export enum Ethnicity {
    /**
     * Human: slavic, scandinavian
     */
    WHITE = 'white',

    /**
     * Human: mongolian
     */
    YELLOW = 'yellow',

    /**
     * Human: guarani, maia
     * Golem: red earth
     */
    RED = 'red',

    /**
     * Human: zulu, xhosa 
     */
    BLACK = 'black',

    /**
     * Human: aboriginal
     * Golem: ground and stone
     */
    BROWN = 'brown',

    /**
     * Golem: dark earth grassed
     */
    GREEN = 'green',

    /**
     * 
     */
    PURPLE = 'purple'
}